scraper/requirements.txt
Flask==3.0.2
boto3==1.35.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dotenv==1.0.1
urllib3==2.1.0

